requirements.txt
torch
torchvision
numpy
pandas
opencv-python-headless
matplotlib
pillow
scikit-learn
timm
segmentation-models-pytorch
grad-cam